% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLstring_to_ambiguity_table.R
\name{GLstring_to_ambiguity_table}
\alias{GLstring_to_ambiguity_table}
\title{GLstring_to_ambiguity_table}
\usage{
GLstring_to_ambiguity_table(GL_string)
}
\arguments{
\item{GL_string}{A GL String that encodes HLA alleles and their potential
ambiguities}
}
\value{
A tibble that contains the expanded GL String with separate columns
for possible gene locations, loci, genotype ambiguities, genotypes, haplotypes,
and alleles, each with associated identifiers
}
\description{
A function that expands a GL String into a longer, more detailed
format (also known as an ambiguity table) by separating the string into its
components resulting from its hierarchical set of operators, including gene
locations, loci, genotypes, haplotypes, and alleles. The function processes
each level of the GL String and assigns identifiers for each hierarchical
component. The resulting table can be assembled back into a GL String using
the function `ambiguity_table_to_GLstring`. This function is an alias of
`GLstring_expand_longer`.
}
\examples{
file <- HLA_typing_1[, -1]
GL_string <- HLA_columns_to_GLstring(file, HLA_typing_columns = everything())
result <- GLstring_to_ambiguity_table(GL_string[1])
print(result)

}
