% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplotype_frequencies-data.R
\docType{data}
\name{Haplotype_frequencies}
\alias{Haplotype_frequencies}
\title{Ten HLA-A~C~B~DRB3/4/5~DRB1~DQA1~DQB1~DPA1~DPB1 haplotypes and their
frequencies in the global population.}
\format{
A tibble:
\describe{
  \item{HLA-A:HLA-DPB1}{HLA allele names}
  \item{Global_Fq}{haplotype frequencies}
  \item{Global_n}{haplotype observations}
  \item{Global_rank}{haplotype rank}
    ...
}
}
\source{
<https://doi.org/10.1016/j.humimm.2021.04.007>
}
\usage{
data(Haplotype_frequencies)
}
\description{
Ten HLA-A~C~B~DRB3/4/5~DRB1~DQA1~DQB1~DPA1~DPB1 haplotypes and their
frequencies in the global population.
}
\keyword{datasets}
