% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_column_repair.R
\name{HLA_column_repair}
\alias{HLA_column_repair}
\title{HLA_column_repair}
\usage{
HLA_column_repair(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame object with column names renamed in the "HLA_A" format.
}
\description{
This function will change column names that have the official
 HLA nomenclature (e.g. "HLA-A*" or "HLA-A") to a format supported by R
 (e.g. "HLA_A"). The dash and asterisk are a special characters in R,
 and makes selecting columns by name difficult.
}
\examples{
data(toydata) # I think we need to make some toydata.
output <- HLA_column_repair(toydata)
}
