% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_dictionary-data.R
\docType{data}
\name{HLA_dictionary}
\alias{HLA_dictionary}
\title{Data on HLA alleles in the 2008 HLA dictionary}
\format{
A tibble:
\describe{
  \item{IPD.Accession}{IPD-IMGT/HLA Accession number}
  \item{HLA.Allele}{HLA allele name}
  \item{WHO.Assigned.Type}{WHO-assigned serologic equivalent}
  ...
}
}
\source{
<https://doi.org/10.1111/j.1399-0039.2008.01183.x>
}
\usage{
data(HLA_dictionary)
}
\description{
Data on HLA alleles in the 2008 HLA dictionary
}
\keyword{datasets}
