% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_mismatch_alleles_HvG.R
\name{HLA_mismatch_alleles_HvG}
\alias{HLA_mismatch_alleles_HvG}
\title{HLA_mismatch_alleles_HvG}
\usage{
HLA_mismatch_alleles_HvG(GL_string_recip, GL_string_donor, locus)
}
\arguments{
\item{GL_string_recip}{GL string for the HLA data of the recipient.}

\item{GL_string_donor}{GL string for the HLA data of the donor.}

\item{locus}{The name of the locus for which the function checks for a mismatch.}
}
\value{
A string of mismatched alleles. If there are no mismatches, returns "NA".
}
\description{
A function that checks for any alleles in the donor that are not present in the recipient for a specified locus and returns a string of mismatched alleles.
}
\examples{
#' test_cases <- tribble(
~case, ~GL_string_recip, ~GL_string_donor, ~locus, ~expected_result,
1, "HLA-A*24:02+HLA-A*29:02", "HLA-A*02:01", "HLA_A", "HLA-A*02:01")

}
