% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_mismatched_alleles_GvH.R
\name{HLA_mismatched_alleles_GvH}
\alias{HLA_mismatched_alleles_GvH}
\title{HLA_mismatch_alleles_HvG}
\usage{
HLA_mismatched_alleles_GvH(
  GL_string_recip,
  GL_string_donor,
  loci,
  homozygous_count = 2
)
}
\arguments{
\item{GL_string_recip}{GL string for the HLA data of the recipient.}

\item{GL_string_donor}{GL string for the HLA data of the donor.}

\item{loci}{The name of the loci for which the function checks for a mismatch; multiple loci should be entered as vectors (e.g.
c("HLA-A", "HLA-DRB1")).}

\item{homozygous_count}{Whether to count homozygous mismatches as 1 or two alleles; 2 is the default.}
}
\value{
A string of mismatched alleles at each locus. If there are no mismatches, returns "NA".
}
\description{
A function that checks for alleles in the recipient that are not present in the donor
for a specified locus and returns a string of mismatched alleles. The function assumes two alleles
at each locus, even if homozygous results are entered as only one allele (e.g. "HLA-A*02:01" and
"HLA-A*02:01+HLA-A*02:01") are both assumed to indicate two identical alleles at the A locus.
If a mismatch from a homozygous locus is detected, two copies of that allele will be output (e.g.
"HLA-A*02:01+HLA-A*02:01"). This behavior can be overridden by setting homozygous_count = 1, so that
these homozygous alleles are only "counted" once. Null alleles (alleles ending in N or n) are not
counted at mismatches.
}
\examples{
GL_string_recip <- "HLA-A*29:02+HLA-A*30:02^HLA-C*06:02+HLA-C*07:01^HLA-B*08:01+HLA-B*13:02^HLA-DRB4*01:03+HLA-DRB4*01:03^HLA-DRB1*04:01+HLA-DRB1*07:01^HLA-DQA1*02:01+HLA-DQA1*03:01^HLA-DQB1*02:02+HLA-DQB1*03:02^HLA-DPA1*01:03+HLA-DPA1*02:01^HLA-DPB1*01:01+HLA-DPB1*16:01"
GL_string_donor <- "HLA-A*03:01+HLA-A*30:01^HLA-C*07:02+HLA-C*12:03^HLA-B*07:02+HLA-B*38:01^HLA-DRB3*01:01^HLA-DRB5*01:01^HLA-DRB1*03:01+HLA-DRB1*15:01^HLA-DQA1*01:02+HLA-DQA1*05:01^HLA-DQB1*02:01+HLA-DQB1*06:02^HLA-DPA1*01:03+HLA-DPA1*01:03^HLA-DPB1*04:01+HLA-DPB1*04:01"
locus <- "A"
result <- HLA_mismatch_alleles_HvG(GL_string_recip, GL_string_donor, locus)
print(result)

}
