% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_typing_1-data.R
\docType{data}
\name{HLA_typing_1}
\alias{HLA_typing_1}
\title{Synthetic HLA typing data for 10 individuals for the HLA-A, B, C, DRB1, DRB3/4/5,
DQB1, DQA1, DPB1 and DPA1 loci.}
\format{
A tibble:
\describe{
  \item{patient}{Patients 1-10}
  \item{A1:DPB1_2}{HLA allele names}
    ...
}
}
\source{
Synthetic data from Nicholas K. Brown.
}
\usage{
data(HLA_typing_1)
}
\description{
Synthetic HLA typing data for 10 individuals for the HLA-A, B, C, DRB1, DRB3/4/5,
DQB1, DQA1, DPB1 and DPA1 loci.
}
\keyword{datasets}
