% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch_table_2010-data.R
\docType{data}
\name{mismatch_table_2010}
\alias{mismatch_table_2010}
\title{Consensus mismatch numbers for every possible allele combination
at a single locus, from the 2010 publication.}
\format{
A tibble:
\describe{
  \item{patient:donor}{allele combinations for patient and donor}
  \item{#GvH:#Max}{consensus mismatch number for each direction}
    ...
}
}
\source{
<https://www.nature.com/articles/bmt2010132>
}
\usage{
data(mismatch_table_2010)
}
\description{
Consensus mismatch numbers for every possible allele combination
at a single locus, from the 2010 publication.
}
\keyword{datasets}
