% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLstring_genes.R
\name{GLstring_genes}
\alias{GLstring_genes}
\title{GLstring_genes}
\usage{
GLstring_genes(.data, gl_string)
}
\arguments{
\item{.data}{A data frame}

\item{gl_string}{The name of the column in the data frame that contains
GL strings}
}
\value{
A data frame with GL strings separated, loci identified, and data
transformed to a wider format with loci as columns.
}
\description{
This function processes a specified column in a data frame
that contains GL strings. It separates the GL strings, identifies the HLA
loci, and transforms the data into a wider format with loci as column names.
}
\examples{
output <- tibble(GL_string = "HLA-A*29:02+HLA-A*30:02^HLA-C*06:02+HLA-C*07:01^HLA-B*08:01+HLA-B*13:02^HLA-DRB4*01:03+HLA-DRB4*01:03^HLA-DRB1*04:01+HLA-DRB1*07:01") \%>\% GLstring_genes("GL_string")
print(output)

}
