% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLstring_genes_expanded.R
\name{GLstring_genes_expanded}
\alias{GLstring_genes_expanded}
\title{GLstring_genes_expanded}
\usage{
GLstring_genes_expanded(data, gl_string)
}
\arguments{
\item{data}{A data frame containing GL strings for HLA data.}

\item{gl_string}{The name of the column in the data frame that contains
GL strings.}
}
\value{
A data frame with expanded columns, where each row has a single
allele for a specific locus.
}
\description{
This function processes a specified column in a data frame
that contains GL strings. It separates the GL strings, identifies the HLA
loci, and transforms the data into a wider format with loci as column names.
It also creates multiple rows to separate each locus in the allele.
}
\examples{
table <- tibble(GL_string = "HLA-A*29:02+HLA-A*30:02^HLA-C*06:02+HLA-C*07:01^HLA-B*08:01+HLA-B*13:02^HLA-DRB4*01:03+HLA-DRB4*01:03^HLA-DRB1*04:01+HLA-DRB1*07:01")

table \%>\% GLstring_genes_expanded("GL_string")

}
