% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLstring_genotype_ambiguity.R
\name{GLstring_genotype_ambiguity}
\alias{GLstring_genotype_ambiguity}
\title{GLstring_genotype_ambiguity}
\usage{
GLstring_genotype_ambiguity(.data, columns, keep_ambiguities = FALSE)
}
\arguments{
\item{.data}{A data frame}

\item{columns}{The names of the columns in the data frame that contain GL strings}

\item{keep_ambiguities}{A logical value indicating whether to retain the
remaining ambiguities in separate columns with "_genotype_ambiguity" appended
to the original column names. Default is FALSE.}
}
\value{
A data frame with the first genotype ambiguity retained in the
original columns. If \code{keep_ambiguities} is TRUE, the remaining
ambiguities are placed in separate columns.
}
\description{
This function processes GL strings in the specified columns of
a data frame to retain only the first genotype ambiguity, optionally
retaining the remaining ambiguities in a separate column with "_ambiguity"
appended. The function ensures that genes have been separated from the GL
strings prior to execution; otherwise, an error will be thrown if a "^" is
detected in the GL strings.
}
\examples{
data(toydata)
output <- GLstring_genotype_ambiguity(toydata, c("GL_column1", "GL_column2"), keep_ambiguities = TRUE)

}
