% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLstring_regex.R
\name{GLstring_regex}
\alias{GLstring_regex}
\title{GLstring_regex}
\usage{
GLstring_regex(data)
}
\arguments{
\item{data}{A string containing an HLA allele.}
}
\value{
A string with the HLA allele formatted as a regex pattern.
}
\description{
This function will format an HLA allele (e.g. "HLA-A*02:01") to
a regex pattern for searching within a GL string. Note that in order for this
function to work properly, the full HLA allele name, including prefixes, is
required. Allele values of "A*02:01" will need to be updated to "HLA-A*02:01",
and "A2" will need to be updated to "HLA-A2". The `HLA_prefix_add` function
is useful in these situations.
}
\examples{
allele <- "HLA-A*02:01"
GLstring_regex(allele)

}
