% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_columns_to_GLstring.R
\name{HLA_columns_to_GLstring}
\alias{HLA_columns_to_GLstring}
\title{HLA_columns_to_GLstring}
\usage{
HLA_columns_to_GLstring(
  data,
  case_column,
  HLA_typing_columns,
  prefix_to_remove = "",
  suffix_to_remove = "",
  serologic = FALSE
)
}
\arguments{
\item{data}{A data frame with each row including an HLA typing result, with
individual columns containing a single allele.}

\item{case_column}{A column that uniquely identifies the typing result.}

\item{HLA_typing_columns}{The list of columns containing the HLA alleles.}

\item{prefix_to_remove}{An optional string of characters to remove from the
locus names.}

\item{suffix_to_remove}{An optional string of characters to remove from the
locus names.}

\item{serologic}{Whether the final type should be formatted in serologic resolution;
molecular resolution is the default.}
}
\value{
A list of GL strings in the order of the original data frame.
}
\description{
A function to take HLA typing data spread across different columns,
as is often found in wild-caught data, and transform it to a GL string.
}
\examples{
data(toydata)
output <- toydata \%>\% mutate(GL_string = HLA_columns_to_GLstring(., patient, c(mA1Cd.recipient:mDPB12cd.recipient), prefix_to_remove = "m", suffix_to_remove = "Cd.recipient"))

}
