% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_columns_to_GLstring.R
\name{HLA_columns_to_GLstring}
\alias{HLA_columns_to_GLstring}
\title{HLA_columns_to_GLstring}
\usage{
HLA_columns_to_GLstring(
  data,
  HLA_typing_columns,
  prefix_to_remove = "",
  suffix_to_remove = "",
  serologic = FALSE
)
}
\arguments{
\item{data}{A data frame with each row including an HLA typing result, with
individual columns containing a single allele.}

\item{HLA_typing_columns}{A list of columns containing the HLA alleles. Tidyselect is supported.}

\item{prefix_to_remove}{An optional string of characters to remove from the
locus names. The goal is to get the column names to the locus and a number. For example,
columns named "mDRB11Cd" and "mDRB12Cd" should use the `prefix_to_remove` value of "m".}

\item{suffix_to_remove}{An optional string of characters to remove from the
locus names. Using the example above, the `suffix_to_remove` value will be "Cd".}

\item{serologic}{Whether the final type should be formatted in serologic resolution;
molecular resolution is the default.}
}
\value{
A list of GL strings in the order of the original data frame.
}
\description{
A function to take HLA typing data spread across different columns,
as is often found in wild-caught data, and transform it to a GL string. If column names
have anything besides the locus name and a number (e.g. "mA1Cd" instead of just "A1"),
the function will have trouble determining the locus from the column name. The `prefix_to_remove`
and `suffix_to_remove` arguments can be used to clean up the column names. See the example for
how these arguments are used.
}
\examples{
typing_table <- tibble(
patient = c("patient1", "patient2", "patient3"),
mA1cd = c("A*01:01", "A*02:01", "A*03:01"),
mA2cd = c("A*11:01", "blank", "A*26:01"),
mB1cd = c("B*07:02", "B*08:01", "B*15:01"),
mB2cd = c("B*44:02", "B*40:01", "-"),
mC1cd = c("C*03:04", "C*04:01", "C*05:01"),
mC2cd = c("C*07:01", "C*07:02", "C*08:01")
)

typing_table \%>\% mutate(GL_string = HLA_columns_to_GLstring(., HLA_typing_columns = c(mA1cd:mC2cd), prefix_to_remove = "m", suffix_to_remove = "cd"))

}
