% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_match_number.R
\name{HLA_match_number}
\alias{HLA_match_number}
\title{HLA_match_number}
\usage{
HLA_match_number(
  GL_string_recip,
  GL_string_donor,
  loci,
  direction = "bidirectional"
)
}
\arguments{
\item{GL_string_recip}{A GL string representing the recipient's HLA genotype.}

\item{GL_string_donor}{A GL string representing the donor's HLA genotype.}

\item{loci}{A character vector specifying the loci to be considered for
mismatch calculation. HLA-DRB3/4/5 (and their serologic equivalents DR51/52/53)
are considered once locus for this function, and should be called in this argument
as "HLA-DRB3/4/5" or "HLA-DR51/52/53", respectively.}

\item{direction}{A character string indicating the direction of match.
Options are "HvG" (host vs. graft), "GvH" (graft vs. host), "bidirectional"
(the minimum value of "HvG" and "GvH").}
}
\value{
An integer value or a character string:
- If `loci` includes only one locus, the function returns an integer
mismatch count for that locus.
- If `loci` includes multiple loci, the function returns a character
string in the format "Locus1=Count1, Locus2=Count2, ...".
}
\description{
Calculates the number of HLA matches as two minus the number of
mismatches from `HLA_mismatch_number`. Homozygous mismatches are counted twice.
Supports match calculations for host-vs-graft (HvG), graft-vs-host (GvH),
or bidirectional. Bidirectional matching is the default, but can be overridden
using the "direction" argument.
}
\examples{
# Example recipient and donor GL strings
GL_string_recip <- "HLA-A*01:01+HLA-A*02:01^HLA-B*07:02+HLA-B*08:01"
GL_string_donor <- "HLA-A*01:01+HLA-A*03:01^HLA-B*07:02+HLA-B*44:02"
loci <- c("HLA-A", "HLA-B")

# Calculate mismatch numbers (Host vs. Graft)
HLA_match_number(GL_string_recip, GL_string_donor, loci, direction = "HvG")

# Calculate mismatch numbers (Graft vs. Host)
HLA_match_number(GL_string_recip, GL_string_donor, loci, direction = "GvH")

# Calculate mismatch numbers (Bidirectional)
HLA_match_number(GL_string_recip, GL_string_donor,
loci, direction = "bidirectional")

}
