% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_match_summary_HCT.R
\name{HLA_match_summary_HCT}
\alias{HLA_match_summary_HCT}
\title{HLA_match_summary_HCT}
\usage{
HLA_match_summary_HCT(
  GL_string_recip,
  GL_string_donor,
  direction = "bidirectional",
  match_grade
)
}
\arguments{
\item{GL_string_recip}{A GL string representing the recipient's HLA genotype,
and minimally containing the HLA-A, B, C and DRB1 loci (for Xof8 matching)
or the HLA-A, B, C, DRB1 and DQB1 loci (for Xof10 matching).}

\item{GL_string_donor}{A GL string representing the donor's HLA genotype,
and minimally containing the HLA-A, B, C and DRB1 loci (for Xof8 matching)
or the HLA-A, B, C, DRB1 and DQB1 loci (for Xof10 matching).}

\item{direction}{"GvH", "HvG" or "bidirectional". Default is "bidirectional".}

\item{match_grade}{"Xof8" for HLA-A, B, C and DRB1 matching or "Xof10" for
HLA-A, B, C, DRB1 and DQB1 matching.}
}
\value{
An integer value of the match grade summary.
}
\description{
Calculates the match summary for either the HLA-A, B, C and DRB1
loci (out-of-8 matching) or the HLA-A, B, C, DRB1 and DQB1 loci (out-of-10 matching),
as is commonly used for hematopoietic cell transplantation (HCT). Homozygous
mismatches are counted twice. Bidirectional matching is the default, but can
be overridden with the "direction" argument.
}
\examples{
# Example recipient and donor GL strings
GL_string_recip <- "HLA-A*01:01+HLA-A*02:01^HLA-B*07:02+HLA-B*08:01"
GL_string_donor <- "HLA-A*01:01+HLA-A*03:01^HLA-B*07:02+HLA-B*44:02"
loci <- c("HLA-A", "HLA-B")

# Calculate mismatch numbers
HLA_match_number(GL_string_recip, GL_string_donor, loci)

}
