% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_mismatch_alleles_HvG.R
\name{HLA_mismatch_alleles_HvG}
\alias{HLA_mismatch_alleles_HvG}
\title{HLA_mismatch_alleles_HvG}
\usage{
HLA_mismatch_alleles_HvG(GL_string_recip, GL_string_donor, locus)
}
\arguments{
\item{GL_string_recip}{GL string for the HLA data of the recipient.}

\item{GL_string_donor}{GL string for the HLA data of the donor.}

\item{locus}{The name of the locus for which the function checks for a mismatch.}
}
\value{
A string of mismatched alleles. If there are no mismatches, returns "NA".
}
\description{
A function that checks for any alleles in the donor that are not present in the recipient for a specified locus and returns a string of mismatched alleles.
}
\examples{
GL_string_recip <- "HLA-A*29:02+HLA-A*30:02^HLA-C*06:02+HLA-C*07:01^HLA-B*08:01+HLA-B*13:02^HLA-DRB4*01:03+HLA-DRB4*01:03^HLA-DRB1*04:01+HLA-DRB1*07:01^HLA-DQA1*02:01+HLA-DQA1*03:01^HLA-DQB1*02:02+HLA-DQB1*03:02^HLA-DPA1*01:03+HLA-DPA1*02:01^HLA-DPB1*01:01+HLA-DPB1*16:01"
GL_string_donor <- "HLA-A*03:01+HLA-A*30:01^HLA-C*07:02+HLA-C*12:03^HLA-B*07:02+HLA-B*38:01^HLA-DRB3*01:01^HLA-DRB5*01:01^HLA-DRB1*03:01+HLA-DRB1*15:01^HLA-DQA1*01:02+HLA-DQA1*05:01^HLA-DQB1*02:01+HLA-DQB1*06:02^HLA-DPA1*01:03+HLA-DPA1*01:03^HLA-DPB1*04:01+HLA-DPB1*04:01"
locus <- "A"
result <- HLA_mismatch_alleles_HvG(GL_string_recip, GL_string_donor, locus)
print(result)

}
