% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_mismatch_base.R
\name{HLA_mismatch_base}
\alias{HLA_mismatch_base}
\title{HLA_mismatch_base}
\usage{
HLA_mismatch_base(
  GL_string_recip,
  GL_string_donor,
  loci,
  direction = c("HvG", "GvH"),
  homozygous_count = 2
)
}
\arguments{
\item{GL_string_recip}{A GL strings representing the recipient's HLA genotypes.}

\item{GL_string_donor}{A GL strings representing the donor's HLA genotypes.}

\item{loci}{A character vector specifying the loci to be considered for
mismatch calculation.}

\item{direction}{A character string indicating the direction of mismatch.
Options are "HvG" (host vs. graft) or "GvH" (graft vs. host).}

\item{homozygous_count}{An integer specifying how to handle homozygosity.
Defaults to 2, where homozygous alleles are treated as duplicated for
mismatch calculations. Can be specified to be 1, in which case homozygous
alleles are treated as single occurrences without duplication.}
}
\value{
A character vector, where each element is a string summarizing the
mismatches for the specified loci. The strings are formatted as
comma-separated locus mismatch entries.
}
\description{
A function to return a string of mismatches between recipient
and donor HLA genotypes represented as GL strings. The function finds
mismatches based on the direction of comparison specified in the inputs
and also handles homozygosity.
}
\examples{
GL_string_recip <- "HLA-A2+HLA-A68^HLA-Cw1+HLA-Cw17^HLA-DR1+HLA-DR17^HLA-DR52
^HLA-DPB1*04:01"
GL_string_donor <- "HLA-A3+HLA-A69^HLA-Cw10+HLA-Cw9^HLA-DR4+HLA-DR17^HLA-DR52
+HLA-DR53^HLA-DPB1*04:01+HLA-DPB1*04:02"
loci <- c("HLA-A", "HLA-Cw", "HLA-DR51/52/53", "HLA-DPB1")
mismatches <- HLA_mismatch_base(GL_string_recip, GL_string_donor, loci, direction = "HvG")
print(mismatches)

# Output
# "HLA-A=HLA-A3+HLA-A69, HLA-Cw=HLA-Cw10+HLA-Cw9, HLA-DR51/52/53=HLA-DR53, HLA-DPB1=HLA-DPB1*04:02"

}
