% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_mismatch_HvG.R
\name{HLA_mismatch_HvG}
\alias{HLA_mismatch_HvG}
\title{HLA_mismatch_HvG}
\usage{
HLA_mismatch_HvG(GL_string_recip, GL_string_donor, loci)
}
\arguments{
\item{GL_string_recip}{GL string for the HLA data of the recipient}

\item{GL_string_donor}{GL string for the HLA data of the donor}

\item{locus}{The name of the locus for which the function checks for a
mismatch}
}
\value{
A boolean indicating whether there is a host vs. graft mismatch
}
\description{
A function that checks whether there are any alleles in the
donor that are not present in the recipient for a specified locus
}
