% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_prefix_add.R
\name{HLA_prefix_add}
\alias{HLA_prefix_add}
\title{HLA_prefix_add}
\usage{
HLA_prefix_add(.data, columns, prefix = "HLA-")
}
\arguments{
\item{.data}{A data frame}

\item{columns}{Name of columns in .data to which the prefix should be added}

\item{prefix}{A character string to be added as a prefix to the column values.
Default is "HLA-".}
}
\value{
A data frame with the specified prefix added to the values in the
selected columns.
}
\description{
This function adds a specified prefix to the beginning of
each value in the identified columns of the given data frame.
}
\examples{
data(toydata)
output <- toydata \%>\% HLA_prefix_add(A.1, "A")

}
