% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_truncate.R
\name{HLA_truncate}
\alias{HLA_truncate}
\title{HLA_truncate}
\usage{
HLA_truncate(data, fields = 2, keep_suffix = TRUE)
}
\arguments{
\item{data}{A string containing an HLA type.}

\item{fields}{An integer specifying the number of fields to retain in the
truncated values. Default is 2.}

\item{keep_suffix}{A logical value indicating whether to retain any
WHO-recognized suffixes. Default is TRUE.}
}
\value{
A string with the HLA typing truncated according to
the specified number of fields and optional suffix retention.
}
\description{
This function truncates the HLA typing values. The truncation
is based on the number of fields specified and optionally retains any
WHO-recognized suffixes (L, S, C, A, Q, or N). Note this will not keep
G or P group designations, as these are defined to specific fields
(3 fields for G groups, and 2 fields for P groups), so if typings
are to be truncated, the G or P group nomenclature is not being preserved.
}
\examples{
typing <- "A*01:01:01:02N"
typing \%>\% HLA_truncate(fields = 2, keep_suffix = FALSE)

}
