% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_truncate.R
\name{HLA_truncate}
\alias{HLA_truncate}
\title{HLA_truncate}
\usage{
HLA_truncate(data, columns, fields = 2, keep_suffix = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{columns}{Names of columns in the data to be truncated}

\item{fields}{An integer specifying the number of fields to retain in the
truncated values. Default is 2.}

\item{keep_suffix}{A logical value indicating whether to retain any
WHO-recognized suffixes. Default is TRUE.}
}
\value{
A data frame with the specified columns truncated according to
the specified number of fields and optional suffix retention.
}
\description{
This function truncates the HLA typing values in the
identified columns of the given data frame. The truncation is based on
the number of fields specified and optionally retains any recognized
suffixes.
}
\examples{
data(toydata)
output <- toydata \%>\% HLA_truncate(A.1, fields = 2, keep_suffix = TRUE)

}
