% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunogenetr-package.R
\docType{package}
\name{immunogenetr-package}
\alias{immunogenetr}
\alias{immunogenetr-package}
\title{immunogenetr: A Comprehensive Toolkit for Clinical HLA Informatics}
\description{
A comprehensive toolkit for clinical Human Leukocyte Antigen (HLA) informatics, built on 'tidyverse' \url{https://tidyverse.tidyverse.org/} principles and making use of genotype list string (GL string, Mack et al. (2023) \doi{10.1111/tan.15126}) for storing and computing HLA genotype data. Specific functionalities include: coercion of HLA data in tabular format to and from GL string; calculation of matching and mismatching in all directions, with multiple output formats; automatic formatting of HLA data for searching within a GL string; truncation of molecular HLA data to a specific number of fields; and reading HLA genotypes in HML files and extracting the GL string. This library is intended for research use. Any application making use of this package in a clinical setting will need to be independently validated according to local regulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/k96nb01/immunogenetr_package}
  \item \doi{10.1016/j.humimm.2025.111619}
  \item Report bugs at \url{https://github.com/k96nb01/immunogenetr_package/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Brown \email{nicholas.brown@pennmedicine.upenn.edu} (\href{https://orcid.org/0000-0002-0046-2315}{ORCID})

Authors:
\itemize{
  \item Busra Coskun \email{bcoskun@sas.upenn.edu} (\href{https://orcid.org/0009-0008-6828-3453}{ORCID})
}

}
\keyword{internal}
