% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HML.R
\name{read_HML}
\alias{read_HML}
\title{read_HML}
\usage{
read_HML(HML_file)
}
\arguments{
\item{HML_file}{The path to an HML file.}
}
\value{
A tibble with the sample name and the GL string.
}
\description{
Reads the GL strings of HML files and returns a tibble with
the full genotype for each sample.
}
\examples{
HML_1 <- system.file("extdata", "HML_1.hml", package = "immunogenetr")
HML_2 <- system.file("extdata", "hml_2.hml", package = "immunogenetr")

read_HML(HML_1)
read_HML(HML_2)

}
